renamed file to fit with names of other helper scripts rename cmake var accordingly
