+++ cmake/create_plugin_xml.py
